+++ framework/packages/abstract-interface/src/account/mod.rs
@@ -13,7 +13,7 @@
 
 use abstract_core::{manager::ModuleInstallConfig, ABSTRACT_EVENT_TYPE};
 
-use crate::{Abstract, AccountDetails, AdapterDeployer};
+use crate::{Abstract, AbstractInterfaceError, AccountDetails, AdapterDeployer};
 
 mod manager;
 mod proxy;
@@ -190,7 +190,6 @@ impl<Chain: CwEnv> AbstractAccount<Chain> {
     {
         self.manager.register_remote_account(host_chain)
     }
-
     pub fn create_sub_account(
         &self,
         account_details: AccountDetails,
@@ -247,12 +246,9 @@ impl<Chain: CwEnv> AbstractAccount<Chain> {
     }
 }
 
-use crate::AbstractInterfaceError;
-impl<T: CwEnv> AbstractAccount<T> {
-    /// Upload and register the account core contracts in the version control if they need to be updated
     pub fn upload_and_register_if_needed(
         &self,
-        version_control: &VersionControl<T>,
+        version_control: &VersionControl<Chain>,
     ) -> Result<bool, AbstractInterfaceError> {
         let mut modules_to_register = Vec::with_capacity(2);
 
@@ -278,5 +274,22 @@ impl<T: CwEnv> AbstractAccount<T> {
         };
 
         Ok(migrated)
+    }
+}
+
+impl<Chain: CwEnv> std::fmt::Display for AbstractAccount<Chain> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "Account manager: {:?} ({:?}) proxy: {:?} ({:?})",
+            self.manager.id(),
+            self.manager
+                .addr_str()
+                .or_else(|_| Result::<_, CwOrchError>::Ok(String::from("unknown"))),
+            self.proxy.id(),
+            self.proxy
+                .addr_str()
+                .or_else(|_| Result::<_, CwOrchError>::Ok(String::from("unknown"))),
+        )
     }
 }