+++ modules/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   
 ### Changed
 
+- Generate message now accepts `sender` instead of `proxy_addr`, allowing use of non-abstract addresses
+
 ### Removed
 
 ### Fixed
+
+- Removed checks for manager in `fetch_data` for dex adapters, to allow using adapter requests by authorized addresses